requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Agent orchestration
langgraph==0.2.50
langchain-core==0.3.19
langsmith==0.1.140

# AI clients
anthropic==0.36.2
google-generativeai==0.8.3

# AWS
boto3==1.35.40
botocore==1.35.40

# Validation
langdetect==1.0.9
tenacity==9.0.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2